Module Arrow_deprecated.Map

val assoc_input : (module Bonsai__Import.Comparator with type comparator_witness = 'cmp and type t = 'key) -> ( 'data, 'result ) t -> ( ( 'key, 'data, 'cmp ) Core.Map.t, ( 'key, 'result, 'cmp ) Core.Map.t ) t

assoc_input is used as a mapping function on a Map.t in order to produce a new map.

val associ_input : (module Bonsai__Import.Comparator with type comparator_witness = 'cmp and type t = 'key) -> ( 'key * 'data, 'result ) t -> ( ( 'key, 'data, 'cmp ) Core.Map.t, ( 'key, 'result, 'cmp ) Core.Map.t ) t

The same as assoc_input but the component has access to the key as well as the data.

val associ_input_with_extra : (module Bonsai__Import.Comparator with type comparator_witness = 'cmp and type t = 'key) -> ( 'key * 'data * 'input, 'result ) t -> ( ( 'key, 'data, 'cmp ) Core.Map.t * 'input, ( 'key, 'result, 'cmp ) Core.Map.t ) t

The same as associ_input but with some extra stuff to be provided to the component in addition to the key and data.