Module Arrow_deprecated.With_incr

val of_incr : 'a Ui_incr.t -> ( _, 'a ) t

Constructs a Bonsai component whose result is always the same as its input Incremental node.

val of_module : (module Bonsai__Import.Component_s_incr with type Action.t = 'a and type Input.t = 'i and type Model.t = 'm and type Result.t = 'r) -> default_model:'m -> ( 'i, 'r ) t

val map : ( 'i, 'r1 ) t -> f:( 'r1 Ui_incr.t -> 'r2 Ui_incr.t ) -> ( 'i, 'r2 ) t

Transforms the result of a component, exposing the incrementality for optimization purposes.

val pure : f:( 'i Ui_incr.t -> 'r Ui_incr.t ) -> ( 'i, 'r ) t

Same as Bonsai.pure but allows the user to optimize using Incremental.

val model_cutoff : ( 'i, 'r ) t -> ( 'i, 'r ) t

Creates a Bonsai component where the given cutoff is applied to the incremental node for the component's model, preventing a component from being recalculated unnecessarily.

See Incr.set_cutoff.

val value_cutoff : cutoff:'i Ui_incr.Cutoff.t -> ( 'i, 'i ) t

Creates a Bonsai component where the given cutoff is applied to the incremental node as input passes through the component, preventing a component from being recalculated unnecessarily.

See Incr.set_cutoff.