Module S.Input

A component receives read-only input, either from the output of other components or from an external system (e.g. data from a server). The input is frequently dynamic, but may also be constant.

type t