Module Cont.Bonsai

include module type of Cont with module For_proc := Cont.For_proc and module Conv := Cont.Conv

module type Enum = Bonsai_private_base.Module_types.Enum

type 'a t

The primary type in the Bonsai library, you can think of a value with type 'a t as "a 'a that changes over time". The two main ways that you get a value with this type are: 1. by creating a state machine

the current value of the state machine is returned as a 'state Bonsai.t

the "send the state machine an action" function is also inside of a Bonsai.t 2. by mapping on existing Bonsai.t's to derive a new computed Bonsai.t

type graph

Bonsai.graph is a required parameter to all Bonsai functions which do more than pure computation. The value is always local_ because Bonsai applications have two phases: 1. The graph building phase. This is the phase where you have access to a local_ Bonsai.graph 2. Runtime. The application has started and modifying the graph is no longer permitted.

Bonsai.t is an applicative, which means that you can let%map on things. In practice, you should prefer let%arr which is conceptually equivalent to let%map, but has some performance benefits.

include Bonsai_private_base.Applicative.S with type 'a t := 'a t

val (<*>) : here:lexing_position -> ('a -> 'b) t -> 'a t -> 'b t

val (<*) : here:lexing_position -> 'a t -> Base.unit t -> 'a t

val (*>) : here:lexing_position -> Base.unit t -> 'a t -> 'a t

val (>>|) : here:lexing_position -> 'a t -> ('a -> 'b) -> 'b t

val apply : here:lexing_position -> ('a -> 'b) t -> 'a t -> 'b t

val all : here:lexing_position -> 'a t Base.list -> 'a Base.list t

val all_unit : here:lexing_position -> Base.unit t Base.list -> Base.unit t

module Applicative_infix : sig ... end

include Bonsai_private_base.Import.Mapn with type 'a t := 'a t

val map3 : here:lexing_position -> 'a1 t -> 'a2 t -> 'a3 t -> f:('a1 -> 'a2 -> 'a3 -> 'b) -> 'b t

val map4 : here:lexing_position -> 'a1 t -> 'a2 t -> 'a3 t -> 'a4 t -> f:('a1 -> 'a2 -> 'a3 -> 'a4 -> 'b) -> 'b t

val map5 : here:lexing_position -> 'a1 t -> 'a2 t -> 'a3 t -> 'a4 t -> 'a5 t -> f:('a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'b) -> 'b t

val map6 : here:lexing_position -> 'a1 t -> 'a2 t -> 'a3 t -> 'a4 t -> 'a5 t -> 'a6 t -> f:('a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'a6 -> 'b) -> 'b t

val map7 : here:lexing_position -> 'a1 t -> 'a2 t -> 'a3 t -> 'a4 t -> 'a5 t -> 'a6 t -> 'a7 t -> f:('a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'a6 -> 'a7 -> 'b) -> 'b t

module Autopack : sig ... end

val return : here:lexing_position -> 'a -> 'a t

return produces a Bonsai.t whose inner value is constant.

val return_lazy : here:lexing_position -> 'a Core.Lazy.t -> 'a t

return_lazy produces a Bonsai.t whose inner value is only computed when it becomes necessary.

val map : here:lexing_position -> 'a t -> f:('a -> 'b) -> 'b t

map, map2, and both are ways to build a new Bonsai.t which is dependent on the values of other Bonsai.t. As noted above, you should prefer to use let%arr than these functions, because they come with some performance benefits.

val map2 : here:lexing_position -> 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t

val both : here:lexing_position -> 'a t -> 'b t -> ('a * 'b) t

val cutoff : here:lexing_position -> 'a t -> equal:('a -> 'a -> bool) -> 'a t

The Bonsai runtime will recompute a node if any of its dependencies change. But sometimes, you may want to consider two contained values to be "close enough" and cut off recomputation. You can do that by passing a custom equality function to Bonsai.cutoff.

val transpose_opt : here:lexing_position -> 'a t option -> 'a option t

transpose_opt flips the order of a Bonsai.t and an option. This is useful for optional args that take Bonsai.ts. Note: the inverse operation is not possible.

val state : here:lexing_position -> ?reset:('model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> 'model -> graph @ local -> 'model t * ('model -> unit Bonsai_private_base.Import.Effect.t) t

Bonsai.state allocates a stateful Bonsai.t node in the graph. It returns both the Bonsai.t containing the current state, as well as a Bonsai.t containing a function for overwriting the state.

You must provide a "starting" value for the state.

?reset is called when the reset effect of a grandparent Bonsai.with_model_resetter is called, and allows you to decide what to do when the model is reset. It defaults to a function that ignores the current model and returns the "starting" value.

?sexp_of_model is only used to give the debugger more information.

?equal (default phys_equal) is used by some combinators to reduce memory usage. (E.g. assoc and match%sub may determine that they can store a reference to only the default model rather than the current model.)

val state_opt : here:lexing_position -> ?reset:('model option -> 'model option) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> ?default_model:'model -> graph @ local -> 'model option t * ('model option -> unit Bonsai_private_base.Import.Effect.t) t

state_opt is just like state except that the model is optional. The model starts out as None unless you provide a value to the default_model optional parameter.

val state' : here:lexing_position -> ?reset:('model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> 'model -> graph @ local -> 'model t * (('model -> 'model) -> unit Bonsai_private_base.Import.Effect.t) t

Similar to state, but the `set` function takes a function that calculates the new state from the previous state.

val toggle : here:lexing_position -> default_model:bool -> graph @ local -> bool t * unit Bonsai_private_base.Import.Effect.t t

Bonsai.toggle is a small helper function for building a bool state that toggles back and forth between true and false whenever the unit Effect.t is scheduled.

module Toggle : sig ... end

val toggle' : here:lexing_position -> default_model:bool -> graph @ local -> Toggle.t

Just like toggle except that you also have an Effect.t for directly setting the state in addition to toggling it back and forth.

module Apply_action_context : sig ... end

A value with the type ('action, 'response) Apply_action_context.t is provided to all state-machine's apply_action functions. It can be used to do a variety of things that are only legal inside of apply_action: 1. Access the application time source directly. This is most likely useful to read the current time or sleep for some time span 2. "inject" a value corresponding to the state-machine's action type into an effect that can be scheduled. 3. Directly schedule effects.

val state_machine : here:lexing_position -> ?reset:(('action, unit) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?sexp_of_action:('action -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> apply_action: (('action, unit) Apply_action_context.t -> 'model -> 'action -> 'model) -> graph @ local -> 'model t * ('action -> unit Bonsai_private_base.Import.Effect.t) t

Bonsai.state_machine allows you to build a state machine whose state is initialized to whatever you pass to default_model, and the state machine transitions states using the apply_action function. The current state and a function for injecting an action into a schedulable effect are returned.

?sexp_of_action is used to provide more information to debugging tools.

?reset, ?sexp_of_model, and ?equal do the same things that they do for Bonsai.state, go read those docs for more.

Previously called state_machine0

module Computation_status : sig ... end

val state_machine_with_input : here:lexing_position -> ?reset:(('action, unit) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?sexp_of_action:('action -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> apply_action: (('action, unit) Apply_action_context.t -> 'input Computation_status.t -> 'model -> 'action -> 'model) -> 'input t -> graph @ local -> 'model t * ('action -> unit Bonsai_private_base.Import.Effect.t) t

A state_machine_with_input is identical to a state_machine except that you can pass an arbitrary 'a Bonsai.t dependency and have access to a 'a Computation_status.t inside of the apply_action function.

In the event that the state-machine is inactive, the computation status will be Inactive

Previously called state_machine1

val actor : here:lexing_position -> ?reset:(('action, 'return) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?sexp_of_action:('action -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> recv: (('action, 'return) Apply_action_context.t -> 'model -> 'action -> 'model * 'return) -> graph @ local -> 'model t * ('action -> 'return Bonsai_private_base.Import.Effect.t) t

Bonsai.actor is similar to Bonsai.state_machine, but its recv function is responsible for not only transitioning the state of the state machine, but also for responding with a "return value" to whoever sent the message to the actor.

val actor_with_input : here:lexing_position -> ?sexp_of_action:('action -> Core.Sexp.t) -> ?reset:(('action, 'return) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> recv: (('action, 'return) Apply_action_context.t -> 'input Computation_status.t -> 'model -> 'action -> 'model * 'return) -> 'input t -> graph @ local -> 'model t * ('action -> 'return Bonsai_private_base.Import.Effect.t) t

actor_with_input is just like actor but it can witness the current value of a Bonsai.t inside its recv function just like state_machine_with_input

module Actor (Action : Core.T1) : sig ... end

Unlike the traditional actor and actor_with_input functions, the Actor functor allows the type of the value returned inside of the effect to change based on which action is being handled.

val freeze : here:lexing_position -> ?sexp_of_model:('a -> Core.Sexp.t) -> ?equal:('a -> 'a -> bool) -> 'a t -> graph @ local -> 'a t

freeze takes a 'a Bonsai.t and returns a new 'a Bonsai.t whose value is frozen to the first value that it witnesses from its input.

val fix : here:lexing_position -> 'input t -> f: (recurse:('input t -> graph @ local -> 'result t) -> 'input t -> graph @ local -> 'result t) -> graph @ local -> 'result t

A fixed-point combinator for writing recursive Bonsai components.

val fix2 : here:lexing_position -> 'a t -> 'b t -> f: (recurse:('a t -> 'b t -> graph @ local -> 'result t) -> 'a t -> 'b t -> graph @ local -> 'result t) -> graph @ local -> 'result t

Just like fix but specialized for 2-arity functions.

val scope_model : here:lexing_position -> ('a, _) Core.Comparator.Module.t -> on:'a t -> for_:(graph @ local -> 'b t) -> graph @ local -> 'b t

scope_model is used to associate the state of some component graph with some other value.

val scope_model_n : here:lexing_position -> ('a, _) Core.Comparator.Module.t -> n:(_, 'b) Autopack.t -> on:'a t -> for_:(graph @ local -> 'b) -> graph @ local -> 'b

Exactly like scope_model but will automatically pack and unpack any number of Bonsai.t produced inside the for_ closure. See documentation for Autopack.t for more info.

val most_recent_some : here:lexing_position -> ?sexp_of_model:('b -> Core.Sexp.t) -> equal:('b -> 'b -> bool) -> 'a t -> f:('a -> 'b option) -> graph @ local -> 'b option t

Bonsai.most_recent_some can be used to find and store a value that has some interesting property by transforming a 'a Bonsai.t into a 'b option Bonsai.t via a f:('a -> 'b option) function. The output of this function is a 'b option Bonsai.t that contains the most recent output of f that returned a Some.

val most_recent_value_satisfying : here:lexing_position -> ?sexp_of_model:('a -> Core.Sexp.t) -> equal:('a -> 'a -> bool) -> 'a t -> condition:('a -> bool) -> graph @ local -> 'a option t

The output of most_recent_value_satisfying is a 'a option Bonsai.t that contains the most recent input value for for which f returned true .

val previous_value : here:lexing_position -> ?sexp_of_model:('a -> Core.Sexp.t) -> equal:('a -> 'a -> bool) -> 'a t -> graph @ local -> 'a option t

previous_value returns a Bonsai.t which contains the value that its input held on the previous frame.

val wrap : here:lexing_position -> ?reset:(('action, unit) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> apply_action: (('action, unit) Apply_action_context.t -> 'result Computation_status.t -> 'model -> 'action -> 'model) -> f: ('model t -> ('action -> unit Bonsai_private_base.Import.Effect.t) t -> graph @ local -> 'result t) -> graph @ local -> 'result t

wrap wraps a Computation (built using f) and provides a model and injection function that the wrapped component can use. Especially of note is that the apply_action for this outer-model has access to the result value of the Computation being wrapped. 'result is wrapped in a Computation_status.t so that you are able to reset the underlying model even if the computation is inactive

val wrap_n : here:lexing_position -> ?reset:(('action, unit) Apply_action_context.t -> 'model -> 'model) -> ?sexp_of_model:('model -> Core.Sexp.t) -> ?equal:('model -> 'model -> bool) -> default_model:'model -> apply_action: (('action, unit) Apply_action_context.t -> 'packed Computation_status.t -> 'model -> 'action -> 'model) -> f: ('model t -> ('action -> unit Bonsai_private_base.Import.Effect.t) t -> graph @ local -> 'unpacked) -> n:('packed, 'unpacked) Autopack.t -> graph @ local -> 'unpacked

Exactly like wrap but will automatically pack and unpack any number of Bonsai.t produced inside the f closure. See documentation for Autopack.t for more info.

val enum : here:lexing_position -> (module Enum with type t = 'k) -> match_:'k t -> with_:('k -> graph @ local -> 'a t) -> graph @ local -> 'a t

enum is used for matching on a value and providing different behaviors on different values. The type of the value must be enumerable (there must be a finite number of possible values), and it must be comparable and sexpable.

The rest of the parameters are named like you might expect from pattern-matching syntax, with match_ taking the value to match on, and with_ taking a function that choose which behavior to use.

val with_inverted_lifecycle_ordering : here:lexing_position -> compute_dep:(graph @ local -> 'dep t) -> f:('dep t -> graph @ local -> 'a t) -> graph @ local -> 'a t

with_inverted_lifecycle_ordering will run compute_dep and pass its result to f, but any lifecycles registered in f will run before those registered in compute_dep.

val with_model_resetter : here:lexing_position -> f:(graph @ local -> 'a t) -> graph @ local -> 'a t * unit Bonsai_private_base.Import.Effect.t t

All stateful components allocated inside of with_model_resetter's f callback will have their reset functions invoked whenever the unit Effect.t returned by with_model_resetter is scheduled.

val with_model_resetter_n : here:lexing_position -> f:(graph @ local -> 'a) -> n:(_, 'a) Autopack.t -> graph @ local -> 'a * unit Bonsai_private_base.Import.Effect.t t

Exactly like with_model_resetter but will automatically pack and unpack any number of Bonsai.t produced inside the f closure. See documentation for Autopack.t for more info.

val with_model_resetter' : here:lexing_position -> f:(reset:unit Bonsai_private_base.Import.Effect.t t -> graph @ local -> 'a t) -> graph @ local -> 'a t

The same as with_model_resetter, but the closure has access to the reset function this time.

val with_model_resetter_n' : here:lexing_position -> f:(reset:unit Bonsai_private_base.Import.Effect.t t -> graph @ local -> 'a) -> n:(_, 'a) Autopack.t -> graph @ local -> 'a

Exactly like with_model_resetter' but will automatically pack and unpack any number of Bonsai.t produced inside the f closure. See documentation for Autopack.t for more info.

val peek : here:lexing_position -> 'a t -> graph @ local -> 'a Computation_status.t Bonsai_private_base.Import.Effect.t t

peek maps a Bonsai.t to an Effect.t with the same underlying value. This allows you to inspect the 'a value from inside of a let%bind.Effect chain which might have been changed by previous effects. It is analogous to peek on other abstract data types, including Deferred.ts and Mvar.ts, but more constrained in that you still can only read from within an effect bind.

The 'a Computation_state.t returned by the effect means that if the value was inactive at the time it is peeked, then the effect will be unable to retrieve it.

module Clock : sig ... end

Various clock and timing functions

module Edge : sig ... end

module Memo : sig ... end

The Memo module can be used to share a computation between multiple components, meaning that if the shared computation is stateful, then the users of that computation will see the same state.

module Effect_throttling : sig ... end

module Dynamic_scope : sig ... end

This module implements dynamic variable scoping. Once a dynamic variable is created, you can store values in it, and lookup those same values. A lookup will find the nearest-most parent "unreverted" set call where a "set" can be "reverted" with set''s revert.

module Incr : sig ... end

val assoc : here:lexing_position -> ('k, 'cmp) Core.Comparator.Module.t -> ('k, 'v, 'cmp) Core.Map.t t -> f:('k t -> 'v t -> graph @ local -> 'a t) -> graph @ local -> ('k, 'a, 'cmp) Core.Map.t t

assoc is used to build a new instance of a Bonsai component for each element of a map.

This function is very similar to Map.mapi or Incr_map.mapi', and for good reason! It is doing the same thing (taking a map and a function and returning a new map with the function applied to every key-value pair), but this function does it with the Bonsai values, which means that the computation is done incrementally and also maintains a state machine for every key-value pair.

val assoc_n : here:lexing_position -> ('k, 'cmp) Core.Comparator.Module.t -> ('k, 'v, 'cmp) Core.Map.t t -> f:('k t -> 'v t -> graph @ local -> 'unpacked) -> n:('packed, 'unpacked) Autopack.t -> graph @ local -> ('k, 'packed, 'cmp) Core.Map.t t

Exactly like assoc but will automatically pack multiple Bonsai.t produced inside the f closure into a tuple. See documentation for Autopack.t for more info.

val all_map : here:lexing_position -> ('k, graph @ local -> 'v t, 'cmp) Core.Map.t -> graph @ local -> ('k, 'v, 'cmp) Core.Map.t t

all_map is like assoc but is only usable when the input map is constant.

val assoc_set : here:lexing_position -> ('key, 'cmp) Core.Comparator.Module.t -> ('key, 'cmp) Core.Set.t t -> f:('key t -> graph @ local -> 'result t) -> graph @ local -> ('key, 'result, 'cmp) Core.Map.t t

Like assoc except that the input value is a Set instead of a Map.

val assoc_set_n : here:lexing_position -> ('key, 'cmp) Core.Comparator.Module.t -> ('key, 'cmp) Core.Set.t t -> f:('key t -> graph @ local -> 'unpacked) -> n:('packed, 'unpacked) Autopack.t -> graph @ local -> ('key, 'packed, 'cmp) Core.Map.t t

Exactly like assoc_set but will automatically pack multiple Bonsai.t produced inside the f closure into a tuple. See documentation for Autopack.t for more info.

val assoc_list : here:lexing_position -> ('key, _) Core.Comparator.Module.t -> 'a list t -> get_key:('a -> 'key) -> f:('key t -> 'a t -> graph @ local -> 'b t) -> graph @ local -> [ `Duplicate_key of 'key | `Ok of 'b list ] t

Like assoc except that the input value is a list instead of a Map. The output list is in the same order as the input list. This function performs O(n log(n)) work (where n is the length of the list) any time that anything in the input list changes, so it may be quite slow with large lists.

val assoc_list_n : here:lexing_position -> ('key, _) Core.Comparator.Module.t -> 'a list t -> get_key:('a -> 'key) -> f:('key t -> 'a t -> graph @ local -> 'unpacked) -> n:('packed, 'unpacked) Autopack.t -> graph @ local -> [ `Duplicate_key of 'key | `Ok of 'packed list ] t

Exactly like assoc_list but will automatically pack multiple Bonsai.t produced inside the f closure into a tuple. See documentation for Autopack.t for more info.

module Time_source = Bonsai_private_base.Import.Time_source

module Debug : sig ... end

module Path : sig ... end

A Bonsai.Path.t is a value that records the path through the bonsai graph that was taken to reach the requesting component. This means that every call to Bonsai.path will yield a distinct and reproducable Bonsai.Path.t, which you can use as keys, or to generate stable unique ids.

val path : here:lexing_position -> graph @ local -> Path.t t

Fetches the current Path.t

val path_id : here:lexing_position -> graph @ local -> string t

Fetches the current Path.t and converts it to a string using Path.to_unique_identifier_string

val arr1 : here:lexing_position -> graph @ local -> ('a t -> (f:('a -> 'b) -> 'b t) @ local) @ local

val arr2 : here:lexing_position -> graph @ local -> ('a t -> ('b t -> (f:('a -> 'b -> 'c) -> 'c t) @ local) @ local) @ local

module Effect = Ui_effect

module Let_syntax : sig ... end

module Map : sig ... end

module Set : sig ... end

val delay : here:lexing_position -> f:(graph @ local -> 'a t) -> graph @ local -> 'a t

Delays the computation in f until it is needed. One use case for this is delaying large match%sub branches that might not initially be in use:

  match%sub x with
  | A -> Bonsai.delay ~f:terrible_computation_1 graph
  | B -> Bonsai.delay ~f:terrible_computation_2 graph
  | C -> Bonsai.delay ~f:terrible_computation_3 graph

Here, if x initially has the value A, then only terrible_computation_1 will be computed at first, while terrible_computation_2 and terrible_computation_3 will be put off until x changes to B or C, respectively.

Note that f will be called at most once.

For recursive components, see fix.

module Expert : sig ... end