Module Poll.Starting

type ('o, 'r) t

val empty : ( 'o, 'o option ) t

empty is an option to pass to the polling functions that changes its return type to be 'o option Computation.t and starting value is None

val initial : 'o -> ( 'o, 'o ) t

initial x is an option to pass to the polling functions that changes its return type to be 'o Computation.t and the starting value is x