Module Let_syntax.Let_syntax

val sub : ?here:Core.Source_code_position.t -> 'a Computation.t -> f:( 'a Value.t -> 'b Computation.t ) -> 'b Computation.t

sub runs a Computation, providing the result of that Computation to the function f in the form of a Value.t. The main way to use this function is via the syntax extension let%sub which is described above.

val switch : match_:int Value.t -> branches:int -> with_:( int -> 'a Computation.t ) -> 'a Computation.t

val map : 'a Value.t -> f:( 'a -> 'b ) -> 'b Value.t

val return : 'a Value.t -> 'a Computation.t

val both : 'a Value.t -> 'b Value.t -> ('a * 'b) Value.t

val arr : ?here:Core.Source_code_position.t -> 'a Value.t -> f:( 'a -> 'b ) -> 'b Computation.t

val map3 : 'a1 Value.t -> 'a2 Value.t -> 'a3 Value.t -> f:( 'a1 -> 'a2 -> 'a3 -> 'b ) -> 'b Value.t

val map4 : 'a1 Value.t -> 'a2 Value.t -> 'a3 Value.t -> 'a4 Value.t -> f:( 'a1 -> 'a2 -> 'a3 -> 'a4 -> 'b ) -> 'b Value.t

val map5 : 'a1 Value.t -> 'a2 Value.t -> 'a3 Value.t -> 'a4 Value.t -> 'a5 Value.t -> f:( 'a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'b ) -> 'b Value.t

val map6 : 'a1 Value.t -> 'a2 Value.t -> 'a3 Value.t -> 'a4 Value.t -> 'a5 Value.t -> 'a6 Value.t -> f:( 'a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'a6 -> 'b ) -> 'b Value.t

val map7 : 'a1 Value.t -> 'a2 Value.t -> 'a3 Value.t -> 'a4 Value.t -> 'a5 Value.t -> 'a6 Value.t -> 'a7 Value.t -> f:( 'a1 -> 'a2 -> 'a3 -> 'a4 -> 'a5 -> 'a6 -> 'a7 -> 'b ) -> 'b Value.t