Module Bonsai.Private

val reveal_value : 'a Value.t -> 'a Bonsai__.Value.t

val conceal_value : 'a Bonsai__.Value.t -> 'a Value.t

val reveal_computation : 'a Computation.t -> 'a Bonsai__.Computation.packed

val conceal_computation : 'a Bonsai__.Computation.packed -> 'a Computation.t

val path : Bonsai__Path.t Computation.t

module Value = Bonsai__.Value

module Computation = Bonsai__.Computation

module Apply_action : sig ... end

module Environment : sig ... end

module Meta : sig ... end

Meta.Model and Meta.Action contain the necessary functions associated with the hidden model and action types of Bonsai components. These are stored in Computation.packed.

module Snapshot : sig ... end

A snapshot represents the state of a component at an instant in time.

module Lifecycle : sig ... end

module Path : sig ... end

module Node_path : sig ... end

module Graph_info : sig ... end

module Instrumentation : sig ... end

module Flatten_values : sig ... end

val eval : environment:Environment.t -> path:Path.t -> clock:Ui_incr.Clock.t -> model:'model Ui_incr.t -> inject_dynamic:( 'dynamic_action -> unit Effect.t ) -> inject_static:( 'static_action -> unit Effect.t ) -> ( 'model, 'dynamic_action, 'static_action, 'result ) Computation.t -> ( 'model, 'dynamic_action, 'result ) Snapshot.t