Module Bonsai_protocol

module Entry : sig ... end

All performance entries are forwarded to the debugger/profiler, whether or not they were generated by Bonsai. Thus, when possible, we extract the Bonsai node that the entry corresponds to, but otherwise we just pass along the label unmodified.

module Message : sig ... end

The debugger/profiler receives both performance entries as well as a description of the Bonsai graph.