Module Bonsai_test.Handle

type ('result, 'incoming) t

val show : ( _, _ ) t -> unit

show prints out the result of the component as specified by the Result_spec that was passed into Handle.create.

val show_diff : ?location_style:Patdiff_kernel.Format.Location_style.t -> ( _, _ ) t -> unit

show_diff will print the diff of the view between now and the last time that show or show_diff was called.

val recompute_view : ( _, _ ) t -> unit

recompute_view is like show, but it doesn't print anything. Calling recompute_view between invocations of show_diff does not affect the diff the gets shown.

val recompute_view_until_stable : ?max_computes:int -> ( _, _ ) t -> unit

This function calls recompute_view until either max_computes is reached (defaults to 100), or there are no more after-display lifecycle events for processing.

This can be useful when using e.g. Bonsai.Edge.on_change, which might otherwise delay their effects until the next frame.

val store_view : ( _, _ ) t -> unit

store_view is like show except that instead of printing the view to stdout, it only stores the current view for use with show_diff. This can be useful if you want to print the diff of "before->after" without being required to print the entirety of "before".

val flush : ( _, _ ) t -> unit

Flushes all the events in the event queue. This can be useful if e.g. you are completing an Effect synchronously via Svar, and don't want to go through the other functions which flush (like show)

val result : ( 'result, _ ) t -> 'result

val do_actions : ( _, 'incoming ) t -> 'incoming list -> unit

val disable_bonsai_path_censoring : ( _, _ ) t -> unit

val disable_bonsai_hash_censoring : ( _, _ ) t -> unit

val clock : ( _, _ ) t -> Ui_incr.Clock.t

val advance_clock_by : ( _, _ ) t -> Core.Time_ns.Span.t -> unit

val create : ( 'result, 'incoming ) Result_spec.t -> ?clock:Ui_incr.Clock.t -> 'result Bonsai.Computation.t -> ( 'result, 'incoming ) t

val show_model : ( _, _ ) t -> unit

val result_incr : ( 'r, 'i ) t -> 'r Ui_incr.t

val lifecycle_incr : ( _, _ ) t -> Ui_incr.Packed.t

val apply_action_incr : ( _, _ ) t -> Ui_incr.Packed.t