Module Effect.Define_visibility

For registering a handler for Viewport_changed events. Note that if this functor is called multiple times, each handler will see all of the events.

Parameters

module VH : Visibility_handler

Signature