Module Bonsai_web_test.Result_spec

include module type of struct include Bonsai_test.Result_spec end

module type S = Bonsai_test.Result_spec.S

type ('result, 'incoming) t = (module S with type incoming = 'incoming and type t = 'result)

module No_incoming = Bonsai_test.Result_spec.No_incoming

include No_incoming is a quick way to define a Result_spec with no incoming events:

module type Sexpable = Bonsai_test.Result_spec.Sexpable

module type Stringable = Bonsai_test.Result_spec.Stringable

val sexp : (module Sexpable with type t = 'a) -> ( 'a, Core.Nothing.t ) t

val string : (module Stringable with type t = 'a) -> ( 'a, Core.Nothing.t ) t

val invisible : ( 'a, Core.Nothing.t ) t

val vdom : ?filter_printed_attributes:( string -> bool ) -> ( 'a -> Bonsai_web.Vdom.Node.t ) -> ( 'a, Core.Nothing.t ) t

filter_printed_attributes controls which attributes on a Node will get printed analyzing the string name of the attribute. Style properties correspond to their property name prefixed with "style.". For example, to filter out the "display" CSS property, you should return false for "style.display"; to filter out all CSS styles, return false when the string begins with "style.". A Node's key corresponds to the string "@key"