Module Bonsai_web_ui_drag_and_drop.For_testing

module Action : sig ... end

Actions for testing use strings instead of the same types in the universe because then the tests do not need a t, but can refer to an existing universe just by their name.

val type_id : ( Action.t -> unit Ui_effect.t ) Core.Type_equal.Id.t