Module For_testing.Action

Actions for testing use strings instead of the same types in the universe because then the tests do not need a t, but can refer to an existing universe just by their name.

type t =

| Start_drag of string

| Set_target of string option

| Finish_drag

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool