Module Bonsai_web_ui_element_size_hooks

module Visibility_tracker : sig ... end

module Size_tracker : sig ... end

module Bulk_size_tracker : sig ... end

module Freeze : sig ... end

Freeze.width/height looks at the size of the node when it enters the dom, discovers its calculated width/height, and then sets it explicitly (in px). This prevents the browser layout from messing with a size that we want to be controlled by the user.

module Resize_to_fit : sig ... end

module Expert : sig ... end