Module Bonsai_web_ui_element_size_hooks.Bulk_size_tracker

module Dimensions : sig ... end

module Options : sig ... end

When attached to a Vdom node, this attribute will monitor the size of this node, and report any changes to the size through the provided callback.

Note: in almost all cases, you'll want to have box-sizing: border-box set on your node in order for it to measure the size of the element by looking at the border-size.

val component : ( 'key, 'cmp ) Bonsai_web.Bonsai.comparator -> 'contained Options.t -> (( 'key, 'contained, 'cmp ) Base.Map.t * ( 'key -> Bonsai_web.Vdom.Attr.t )) Bonsai_web.Computation.t

module For_testing : sig ... end