Module Bulk_size_tracker.Options

type 'a maybe_stale =

| Fresh of 'a

| Stale of 'a

val sexp_of_maybe_stale : ( 'a -> Sexplib0.Sexp.t ) -> 'a maybe_stale -> Sexplib0.Sexp.t

val maybe_stale_of_sexp : ( Sexplib0.Sexp.t -> 'a ) -> Sexplib0.Sexp.t -> 'a maybe_stale

val equal_maybe_stale : ( 'a -> 'a -> bool ) -> 'a maybe_stale -> 'a maybe_stale -> bool

type 'a t =

| Prune_stale : Dimensions.t t

| Keep_stale : Dimensions.t maybe_stale t

| Ignore_stale : Dimensions.t t

Prune_stale will remove all the items from the output map whenever the element is removed from the DOM, while Keep_stale will keep the previous values around, but mark them as being stale.