Module Bonsai_web_ui_element_size_hooks.Freeze

Freeze.width/height looks at the size of the node when it enters the dom, discovers its calculated width/height, and then sets it explicitly (in px). This prevents the browser layout from messing with a size that we want to be controlled by the user.

val width : Bonsai_web.Vdom.Attr.t

val height : Bonsai_web.Vdom.Attr.t

module Expert : sig ... end