Module Bonsai_web_ui_element_size_hooks.Size_tracker

val on_change : ( width:float -> height:float -> unit Ui_effect.t ) -> Bonsai_web.Vdom.Attr.t

When attached to a Vdom node, this attribute will monitor the size of this node, and report any changes to the size through the provided callback.

Note: in almost all cases, you'll want to have box-sizing: border-box set on your node in order for it to measure the size of the element by looking at the border-size.

module For_testing : sig ... end