Module Bonsai_web_ui_extendy

module Id : sig ... end

type 'a t = {

contents : 'a Id.Map.t;

append : unit Ui_effect.t;

set_length : int -> unit Ui_effect.t;

remove : Id.t -> unit Ui_effect.t;

}

val component : Core.Source_code_position.t -> 'a Bonsai.Computation.t -> 'a t Bonsai.Computation.t

Given a computation, builds a new computation that can hold a dynamic number of the wrapped computation.

val component' : Core.Source_code_position.t -> 'a Bonsai.Computation.t -> wrap_remove:( 'a -> unit Ui_effect.t -> 'b ) -> 'b t Bonsai.Computation.t

Like component, but with the power to extend the result of the input component with an event that removes itself.