Module Bonsai_web_ui_file.For_testing

This module supports a test / simulation mode where file data can be fed in by you rather than coming from an actual file on disk.

module Test_data : sig ... end

val create : Test_data.t -> t