Module Elements.File_select

A form element that allows the user to select a file from their local disk.

NOTE: these widgets are not safe for use in Tangle as internally they require a model which cannot be of_sexp'd.

val single_opt : Core.Source_code_position.t -> ?extra_attrs:Bonsai_web.Vdom.Attr.t list Bonsai_web.Value.t -> ?accept:[ `Extension of string | `Mimetype of string ] list -> unit -> Bonsai_web_ui_file.t option Bonsai_web_ui_form__Form.t Bonsai_web.Computation.t

val single : Core.Source_code_position.t -> ?extra_attrs:Bonsai_web.Vdom.Attr.t list Bonsai_web.Value.t -> ?accept:[ `Extension of string | `Mimetype of string ] list -> unit -> Bonsai_web_ui_file.t Bonsai_web_ui_form__Form.t Bonsai_web.Computation.t

A form where picking a file is mandatory. The form will be in an error state until a file is picked.

val multiple : Core.Source_code_position.t -> ?extra_attrs:Bonsai_web.Vdom.Attr.t list Bonsai_web.Value.t -> ?accept:[ `Extension of string | `Mimetype of string ] list -> unit -> Bonsai_web_ui_file.t Core.Filename.Map.t Bonsai_web_ui_form__Form.t Bonsai_web.Computation.t