Module Bonsai_web_ui_form.Submit

type 'a t

Creates a "submit" handler, which is intended to be used by the view_as_vdom function.

handle_enter: when true, will render the form inside of a <form> element, which gives us the ability to add an "on_submit" handler which detecs people hitting "enter" when a form element is focused. The default is true. Set it to false to disable this behavior.

button: When Some, will append a button with the label given by its contents. will be added to the end of the form. If the form is currently invalid, the button will be disabled. The default is Some "Submit". Explicitly set it to None to remove the button entirely.

f: the function which is run when the form is submitted.

val create : ?handle_enter:bool -> ?button:string option -> f:( 'a -> unit Ui_effect.t ) -> unit -> 'a t