Module type Record.S

module Typed_field : Typed_fields_lib.S

val form_for_field : 'a Typed_field.t -> 'a Bonsai_web_ui_form__Form.t Bonsai_web.Computation.t

For each of the fields in your record, you need to provide a form component which produces values of that type.