Module type Variant.S

module Typed_variant : Typed_variants_lib.S

This module should be generated by deriving typed_variants on a sum type.

val form_for_variant : 'a Typed_variant.t -> 'a Bonsai_web_ui_form__Form.t Bonsai_web.Computation.t

For each of the variants in your sum type, you need to provide a form component which produces values of that type.