Module Bonsai_web_ui_form.View

Bonsai-web-forms has its own view type so that it can build a more structured view of the form. You can convert it to a Vdom node with the to_vdom* functions.

module Private = View

type t = Private.t

val to_vdom : ?custom: ( ?on_submit:Private.submission_options -> ?editable:Private.editable -> t -> Bonsai_web.Vdom.Node.t ) -> ?on_submit:Private.submission_options -> ?editable:Private.editable -> t -> Bonsai_web.Vdom.Node.t

Converts the form to a vdom node. By default, a table. A custom converter (likely written using the View.Expert module) can be specified to be used instead.

val to_vdom_plain : t -> Bonsai_web.Vdom.Node.t list

This function can be used to acquire the full list of leaf-node form elements sans any formatting, labeling, or error reporting.