Module Bonsai_web_ui_multi_select.Make

Parameters

module Item : Item

Signature

A widget for selecting multiple items from a list. A fairly rich interface is provided, including:

Using the arrow keys and Enter to select / deselect items;

The ability to filter the set of items displayed by searching;

"Select all" and "Select none" buttons

module View_config : sig ... end

module Selection_status : sig ... end

module Action : sig ... end

module Result : sig ... end

module Initial_model_settings : sig ... end

val bonsai : ?initial_model_settings:Initial_model_settings.t -> ?default_selection_status:Selection_status.t Bonsai_web.Value.t -> view_config:View_config.t Bonsai_web.Value.t -> Item.Set.t Bonsai_web.Value.t -> Result.t Bonsai_web.Computation.t