Module Bonsai_web_ui_not_connected_warning_box

val message_for_async_durable : Core.Time_ns.Span.t -> string

val component : ?styles:Virtual_dom.Vdom.Attr.t -> create_message:( Core.Time_ns.Span.t -> string ) -> bool Bonsai.Value.t -> Virtual_dom.Vdom.Node.t Bonsai.Computation.t

A warning box which appears in the bottom-right corner of a web page if the input is true, and behaves like Vdom.Node.none if the input is false.

This component comes with some default styles, but those may be overriden. The create_message function is given the length of time that the input value has been false (i.e. how long the server has been disconnected).

If you are using async_durable, we recommend defaulting to message_for_async_durable.