Module Bonsai_web_ui_notifications

This is a notifications component that provides pretty and user-friendly(ish) notifications, with automatic fade-out & dismissal after a time span.

type t

val create : ?dismiss_notifications_after:Core.Time_ns.Span.t Bonsai_web.Value.t -> ?dismiss_errors_automatically:bool Bonsai_web.Value.t -> Core.Source_code_position.t -> t Bonsai_web.Computation.t

val add_error : ?error:Core.Error.t -> t -> text:string -> unit Ui_effect.t

add_error will create a notification with text in primary focus and error pretty printed in small font (if provided).

val add_success : t -> text:string -> unit Ui_effect.t

add_success creates a notification with only text in primary focus.

module Notification_style : sig ... end

val to_vdom : ?notification_style:Notification_style.t -> ?notification_extra_attr:Bonsai_web.Vdom.Attr.t -> t -> Bonsai_web.Vdom.Node.t

to_vdom constructs the vdom necessary to display these notifications. The styling required to display these notifications is handled for you via ppx_css.