Module Bonsai_web_ui_notifications.Notification_style

module type S = sig ... end

These CSS classes are applied to the notifications that are generated by to_vdom. To customise the color of these notifications, set the background-color attribute to your desired color.

type t = (module S)