Module Bonsai_web_ui_partial_render_table.Basic

module Focus = Focus

module Result : sig ... end

module Columns : sig ... end

There are a few ways to specify the columns on a partial render, table, and they each have their own tradeoffs and capibilities. You can not mix-and-match column kinds. Read the doc comments for each of the submodules to learn more.

val component : ?filter:( key:'key -> data:'data -> bool ) Bonsai_web.Value.t -> ?default_sort:( ('key * 'data) -> ('key * 'data) -> int ) Bonsai_web.Value.t -> ?preload_rows:int -> ( 'key, 'cmp ) Bonsai_web.Bonsai.comparator -> focus:( 'focus, 'key ) Focus.t -> row_height:[ `Px of int ] -> columns:( 'key, 'data ) Columns.t -> ( 'key, 'data, 'cmp ) Core.Map.t Bonsai_web.Value.t -> 'focus Result.t Bonsai_web.Computation.t

This is the main UI component for the table content.