Module Columns.Dynamic_cells

Dynamic_cells is a column-specification format with the following tradeoffs:

Pro: Each cell is it's own bonsai computation, so you can stick complex components in side of them, like forms, or graphs.

Con: The set of columns must be statically known ahead of time, and can not be determined dynamically.

type ('key, 'data) t

val column : ?sort:( ('key * 'data) -> ('key * 'data) -> int ) Bonsai_web.Value.t -> ?initial_width:[ `Px of int ] -> ?visible:bool Bonsai_web.Value.t -> label:Bonsai_web.Vdom.Node.t Bonsai_web.Value.t -> cell: ( key:'key Bonsai_web.Value.t -> data:'data Bonsai_web.Value.t -> Bonsai_web.Vdom.Node.t Bonsai_web.Computation.t ) -> unit -> ( 'key, 'data ) t

val group : label:Bonsai_web.Vdom.Node.t Bonsai_web.Value.t -> ( 'key, 'data ) t list -> ( 'key, 'data ) t

group ~label children builds a header-group that has children underneath it. The content of header-group is set to label

val expand : label:Bonsai_web.Vdom.Node.t Bonsai_web.Value.t -> ( 'key, 'data ) t -> ( 'key, 'data ) t

expand ~label child builds a header-group that has a single child underneath it.

val lift : ( 'key, 'data ) t list -> ( 'key, 'data ) columns

lift pulls a list of columns out into a column specification for use in the primary APIs