Module Columns.Dynamic_columns

type ('key, 'data) t

Dynamic_columns is a column-specification format with the following tradeoffs:

Pro: The set of columns, and how to render them can be determined dynamically (lift takes a column list inside a Value.t)

Con: Cells are computed with plain functions, and can not maintain state.

val column : ?sort:( ('key * 'data) -> ('key * 'data) -> int ) -> ?initial_width:[ `Px of int ] -> ?visible:bool -> label:Bonsai_web.Vdom.Node.t -> cell:( key:'key -> data:'data -> Bonsai_web.Vdom.Node.t ) -> unit -> ( 'key, 'data ) t

val group : label:Bonsai_web.Vdom.Node.t -> ( 'key, 'data ) t list -> ( 'key, 'data ) t

group ~label children builds a header-group that has children underneath it. The content of header-group is set to label

val lift : ( 'key, 'data ) t list Bonsai_web.Value.t -> ( 'key, 'data ) columns

lift pulls a list of columns out into a column specification for use in the primary APIs