Module Bonsai_web_ui_search_bar

This library provides an alternative to an HTML select. Notable features :

Allows users to type in text to filter results

Bounded number of DOM appends per search

Autocomplete can display the choice as well as associated metadata

module type Item = sig ... end

module Input : sig ... end

val create : (module Item with type t = 'item) -> ?max_query_results:int -> ?width:Css_gen.Length.t -> ?placeholder:string -> ?initial_query:string -> ?wrap_search_bar:( Bonsai_web.Vdom.Node.t -> Bonsai_web.Vdom.Node.t ) -> ?autocomplete_item:( 'item -> Bonsai_web.Vdom.Node.t ) -> ?filter_choice:( input_text:string -> 'item -> bool ) -> ?score_choice:( input_text:string -> 'item -> int ) -> ?of_string:( string -> 'item option ) -> ?extra_textbox_attrs:Bonsai_web.Vdom.Attr.t list -> unit -> 'item Input.t Bonsai_web.Value.t -> Bonsai_web.Vdom.Node.t Bonsai_web.Computation.t