Module Bonsai_web_ui_tabs

module State : sig ... end

module Result : sig ... end

val tab_state : (module Bonsai_web.Bonsai.Model with type t = 'a) -> initial:'a -> 'a State.t Bonsai_web.Computation.t

tab_state is used to build a basic state that can be used to store the state for a current tab.

val tab_ui : ?decorate:( 'a -> Bonsai_web.Vdom.Node.t ) Bonsai_web.Value.t -> ?additional_button_attributes: ( is_selected:bool -> 'a -> Bonsai_web.Vdom.Attr.t list ) Bonsai_web.Value.t -> (module Bonsai_web.Bonsai.Model with type t = 'a) -> all_tabs:'a list Bonsai_web.Value.t -> 'a State.t Bonsai_web.Value.t -> f: ( change_tab:( 'a -> unit Bonsai_web.Vdom.Effect.t ) Bonsai_web.Value.t -> 'a Bonsai_web.Value.t -> 'result Bonsai_web.Computation.t ) -> 'result Result.t Bonsai_web.Computation.t

tab_ui takes a state and evaluates a computation-returning function that can match on the tab and produce a computation for that tab.

decorate can be used to produce the contents for the button for each tab.

additional_button_attributes can be used to add more Vdom.Attr.ts to the button nodes for each tab.