Module Bonsai_web_ui_tabs.Result

type 'a t = {

tabs : Bonsai_web.Vdom.Node.t list;

current : 'a;

}

The result of a tabs UI component is the list of tabs, and the result from evaluating the current tab.

val combine_trivially : Bonsai_web.Vdom.Node.t t -> Bonsai_web.Vdom.Node.t

When the result from the tabs component is just a Vdom.Node.t, there's an easy way to combine them. This will make a Vdom node that looks like this:

        <div class="bonsai_ui_tab_container">
            <div class="bonsai_ui_tab_tabs"> ... tabs here ... </div>
            <div class="bonsai_ui_tab_body"> ... current tab here ... </div>
        </div>