Module Bonsai_web_ui_tabs.State

type 'a t

val current : 'a t -> 'a

val set : 'a t -> 'a -> unit Ui_effect.t

val create : current:'a -> set:( 'a -> unit Ui_effect.t ) -> 'a t

The primary way to create a State.t is to use tab_state below. create should only be used if there is an external source that drives the tab select, such as a URL.