Module Bonsai_web_ui_typeahead

module Styles : sig ... end

This file exists to encapsulate all the styles used within these form elements. Ideally these form components would be web components https://developer.mozilla.org/en-US/docs/Web/Web_Components and styles would be encapsulated within a shadow dom. This doesn't appear to be supported in jsoo. (at least the bindings don't exist)

module Typeahead : sig ... end

These controls come unstyled by default. jane-web-style provides css that will make the control and option pills pretty.